% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_CIs.R
\name{bootstrap_pffr}
\alias{bootstrap_pffr}
\title{Perform bootstrap for smooth effects in \code{\link[refund]{pffr}} models}
\usage{
bootstrap_pffr(type = "parametric", B = 1000, formula, data, model = NULL,
  param_yvar = NULL, cores = 1, param_simFun = "simulate",
  param_yMinValue = NULL, log_file = NULL, ...)
}
\arguments{
\item{type}{One of \code{"parametric"} or \code{"nonparametric"}, specifying
the type of bootstrap to be used. If \code{type = "parametric"} you also have
to specify the \code{model} argument.}

\item{B}{Number of bootstrap samples, defaults to 1000}

\item{formula}{See formula argument of \code{\link[refund]{pffr}}}

\item{data}{See data argument of \code{pffr}}

\item{model}{Function-on-scalar model fitted with \code{pffr} used for
parametric bootstrapping. Only used if \code{type=='parametric'}.}

\item{param_yvar}{Name of response variable used in \code{model} and present
in \code{data}. Only used if \code{type=='parametric'}.}

\item{cores}{Number of cores to use for parallel processing. Possible for
both Linux-based systems and Windows.}

\item{param_simFun}{Specifies the simulation function to be used for
parametric bootstrapping. See \code{\link{simulate_pffr}}.}

\item{param_yMinValue}{Minimum value to which the bootstrapped response values
are set using a parametric bootstrap. See \code{\link{simulate_pffr}}.}

\item{log_file}{Optional filename of a log file where progress of parallel
processing should printed to. Defaults to console output (only on Linux-based
systems). Only used if \code{cores>1}.}

\item{...}{Further arguments passed to \code{pffr}}
}
\description{
With this function you can perform a parametric or nonparametric
bootstrap for a function-on-scalar model fitted with \code{\link[refund]{pffr}}.
Based on the results, confidence intervals (CIs) for smooth effects can be
calculated using \code{link{calc_bootstrapCIs}}.
}
