% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_CIs.R
\name{calc_bootstrapCIs}
\alias{calc_bootstrapCIs}
\title{Calculate bootstrap-based confidence intervals for smooth effects in \code{\link[refund]{pffr}} models}
\usage{
calc_bootstrapCIs(bs_results, alpha = 0.05)
}
\arguments{
\item{bs_results}{Output of function \code{\link{bootstrap_pffr}}}

\item{alpha}{\code{(1-alpha)} CIs are calculated. The default 0.05 leads to 95\% CIs}
}
\description{
Wrapper for \code{\link{calc_singleBootstrapCI}} that calculates CIs for all
smooth terms in \code{bs_results}.
}
