% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plot_helpers.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Create a grid of multiple ggplots sharing the same legend}
\usage{
grid_arrange_shared_legend(..., nrow = 1, ncol = length(list(...)),
  legend_position = c("bottom", "right"), main = NULL, sub = NULL,
  outer_fontsize = 15, newpage = TRUE)
}
\arguments{
\item{...}{Multiple \code{\link[ggplot2]{ggplot}} objects}

\item{nrow, ncol}{Parameters specifying the grid to arrange the ggplots in.
Defaults to one row and multiple columns for the plots.}

\item{legend_position}{The legend can be plotted at the \code{"bottom"} or on
the \code{"right"} of all plots}

\item{main, sub}{Text over and under the whole graphic, respectively}

\item{outer_fontsize}{Fontsize of \code{main} and \code{sub}. Defaults to 15.}

\item{newpage}{If TRUE, \code{\link[grid]{grid.newpage}} is called before
plot creation. This is necessary for the plot window in R, but should be
suppressed if one exports the plot, e.g. using \code{\link{pdf}}}
}
\value{
Silently returns the plot object
}
\description{
Function to create a grid of multiple \code{\link[ggplot2]{ggplot}}s sharing
the same legend.
}
