% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plot_helpers.R
\name{no.plot}
\alias{no.plot}
\title{Extract returned values of plot.gam() while suppressing creation of the plot}
\usage{
no.plot(model)
}
\arguments{
\item{model}{Function-on-scalar model fitted with \code{\link[refund]{pffr}}}
}
\description{
Function to extract the values returned of \code{\link[refund]{plot.pffr}} while
suppressing creation of the plot.
}
