% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plot_helpers.R
\name{pi_gamma}
\alias{pi_gamma}
\title{Estimate posterior simulation prediction intervals for Gamma regression}
\usage{
pi_gamma(model, newdata, alpha, n.rep = 10000, seed = NULL)
}
\arguments{
\item{model}{Function-on-scalar regression model fitted with \code{\link[refund]{pffr}}}

\item{newdata}{The new data.frame for which the predictions are calculated}

\item{alpha}{Alpha level, defaults to 0.05, i.e. 95\% prediction intervals}

\item{n.rep}{number of replications for the posterior simulation}

\item{seed}{Optional seed for reproducibility}
}
\value{
\code{data.frame} with two columns \code{pi_lower} and \code{pi_upper}
}
\description{
Prediction intervals for a function-on-scalar Gamma regression model
are estimated using the 'posterior simulation' approach for the parameter vector,
e.g. outlined here: https://stat.ethz.ch/pipermail/r-help/2011-April/275632.html.
This approach is more generally applicable to generalized regression models,
however the implementation is currently only focused on Gamma regression.
}
