% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plots.R
\name{plot_2D}
\alias{plot_2D}
\title{3D perspective plot for 2D smooth effects for \code{\link[refund]{pffr}} models}
\usage{
plot_2D(model, select, theta = 30, phi = 30, ...)
}
\arguments{
\item{model}{Function-on-scalar model fitted with \code{\link[refund]{pffr}}}

\item{select}{Index of the smooth effect to be plotted}

\item{theta, phi}{Some specific arguments for \code{\link{persp}} with
changed default values}

\item{...}{Further arguments passed to \code{persp}}
}
\description{
This function produces 3D plots for 2D smooth effects of function-on-scalar 
models fitted with \code{\link[refund]{pffr}}.
This function mainly does the same as \code{mgcv::plot.gam(..., pers = TRUE)}.
However, \code{\link[mgcv]{plot.gam}} only uses the \code{shade} argument
for 1D smooths and doesn't pass it to the \code{\link{persp}} function. Such
arguments not passed to \code{persp} can be used with this function.
}
