% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plots.R
\name{plot_2Dheatmap}
\alias{plot_2Dheatmap}
\title{Heatmat plot for 2D smooth effects for \code{\link[refund]{pffr}} models}
\usage{
plot_2Dheatmap(model, plot_type = 1, plot_ci = TRUE, plot_ci_type = "all",
  alpha = 0.05, CIs_list, select, lowCol = "blue", highCol = "red",
  legend_limits, legend_position = "right", xlab, ylab, main, sub = NULL,
  effect_label, base_size = 11, outer_fontsize = 15, newpage = TRUE, ...)
}
\arguments{
\item{model}{Function-on-scalar model fitted with \code{\link[refund]{pffr}}}

\item{plot_type}{Either 1 (plotting CIs after Marra \& Wood (2012)) or 2
(bootstrap-based CIs, by specyfing \code{CIs_list}). Defaults to 1.}

\item{plot_ci}{If \code{TRUE} CIs are plotted. Only used if \code{plot_type = 1}.}

\item{plot_ci_type}{Only used if \code{plot_ci = TRUE}. One of
\code{c('all','lower','upper')}. Determines if the effect and the CI
boundaries (\code{'all'}), only the lower or only the upper boundary
(\code{'lower'/'upper'}) should be plotted}

\item{alpha}{\code{(1-alpha)} CIs are calculated. The default 0.05 leads to 
95\% CIs. Only used if \code{plot_type = 1}.}

\item{CIs_list}{Output of function \code{\link{calc_bootstrapCIs}}.
Only necessary if \code{plot_type = 2}.}

\item{select}{Index of smooth term to be plotted. If a vector of length two
is passed, then the difference of the second smooth minus the first smooth will
be plotted. Note that CIs from \code{plot_type == 1} are suppressed in this case.
However, using the \code{CIs_list} argument one can still plot boostrapped
CIs for such a difference of smooths, by passing a list with only one CI element.
Such a CI element can be obtained by using \code{\link{calc_singleBootstrapCI}},
using the \code{select} argument.}

\item{lowCol, highCol}{Colors defining the color gradient of the plots}

\item{legend_limits}{Optional vector specifying the limits of the color
scale, e.g. \code{c(-2,2)}}

\item{legend_position}{Either "bottom", "right" or "none"}

\item{xlab}{Optional parameters for overwriting the default
plot parameters}

\item{ylab}{Optional parameters for overwriting the default
plot parameters}

\item{main}{Optional parameters for overwriting the default
plot parameters}

\item{sub}{Optional subtitle of the plot}

\item{effect_label}{Used in the main title of the plot as
'\code{effect_label} with global 95\% CI'. Only used if \code{main} is not specified.}

\item{base_size}{Size of plot elements. see \code{\link[ggplot2]{theme_bw}}.
Defaults to 11.}

\item{outer_fontsize}{Fontsize of the main and sub which are outer of the
single ggplots. Defaults to 15.}

\item{newpage}{See \code{\link[FoSIntro]{grid_arrange_shared_legend}} function.
Should be set to FALSE when exporting the plot, e.g. using \code{\link{pdf}}.}

\item{...}{Further arguments passed to \code{\link[ggplot2]{theme}}}
}
\value{
The function silently returns the ggplot object
}
\description{
Plots heatmap plots for 2D smooth effects for a function-on-scalar model
fitted with \code{\link[refund]{pffr}}. Additionally, Marra \& Wood (2012) or
bootstrapped-based confidence intervals (CIs) can be plotted, specifying the
\code{plot_type} argument.
}
