% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_CIs.R
\name{plot_bootstrapCIs}
\alias{plot_bootstrapCIs}
\title{Plot bootstrapped confidence intervals}
\usage{
plot_bootstrapCIs(model, CIs_list, select = NULL, xlab, ylab, main,
  effect_label, ...)
}
\arguments{
\item{model}{Function-on-scalar model fitted with \code{\link[refund]{pffr}},
which was fit on the raw, unbootstrapped data to extract point estimate.}

\item{CIs_list}{Output of function \code{\link{calc_bootstrapCIs}}}

\item{select}{Index of smooth term to be plotted. Default plots all effects}

\item{xlab, ylab, main}{Optional parameters to overwrite the default
plot parameters}

\item{effect_label}{Used in main as '\code{effect_label} with global 95\% CI'.
Only used if \code{main = NULL}.}

\item{...}{Further arguments passed to \code{\link{plot_2Dheatmap}}}
}
\value{
Silently returns list with plot objects
}
\description{
Function to plot bootstrapped confidence intervals for function-on-scalar
models fitted with \code{\link[refund]{pffr}}. The bootstrap confidence intervals
have to be precalculated beforehand using \code{\link{calc_bootstrapCIs}}.
}
