% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plots.R
\name{plot_predVSobs}
\alias{plot_predVSobs}
\title{Plot predictions vs real observations for \code{\link[refund]{pffr}} models}
\usage{
plot_predVSobs(model, data, yvar, yvar_label, type = "response",
  log10 = FALSE, plot_pi = FALSE, pi_alpha = 0.05, base_size = 11,
  xlab = "yindex", ylab, main = "Prediction vs observation", ybreaks,
  ylabels, ylim = NULL, lwd = 1, legend_symbol_size = 2)
}
\arguments{
\item{model}{Function-on-scalar model fitted with \code{\link[refund]{pffr}}}

\item{data}{Dataset with one row for which the predictions are to be plotted}

\item{yvar}{name of the response variable in \code{model}}

\item{yvar_label}{Optional label of the response variable used for plot annotation}

\item{type}{\code{"response"} (default) or \code{"link"}.
See \code{\link[refund]{predict.pffr}}.}

\item{log10}{If TRUE the predictions are shown on a log10 scale}

\item{plot_pi}{If \code{TRUE} prediction intervals are plotted.}

\item{pi_alpha}{If \code{plot_pi} is \code{TRUE}, \code{1-pi_alpha}\% prediction
intervals are plotted. Defaults to 0.05.}

\item{base_size}{Base size of plot elements, see \code{\link[ggplot2]{theme_bw}}.
Defaults to 11.}

\item{xlab, ylab, main}{Optional arguments for plot annotation}

\item{ybreaks, ylabels}{Optional specification of the tick position and the
corresponding labels (character vector) for the y axis}

\item{ylim}{Optional y axis limits of plot}

\item{lwd}{Optional line width. Defaults to 2}

\item{legend_symbol_size}{Optional size of symbols in legend.}
}
\description{
Plots the predictions o a function-on-scalar model fitted with
\code{\link[refund]{pffr}} agains the real observations.
}
