% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plots.R
\name{plot_residAutocov}
\alias{plot_residAutocov}
\title{Residual autocovariance plot for \code{\link[refund]{pffr}} models}
\usage{
plot_residAutocov(model, base_size = 11, legend.position = "bottom",
  legend.key.width = unit(2, "lines"), cov.use = "everything", ...)
}
\arguments{
\item{model}{Function-on-scalar model fitted with \code{\link[refund]{pffr}}}

\item{base_size}{Size of plot elements. see \code{\link[ggplot2]{theme_bw}}.
Defaults to 11.}

\item{legend.position, legend.key.width}{Arguments for \code{\link[ggplot2]{theme}}}

\item{cov.use}{see \code{use} argument in \code{\link[stats]{cov}}}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{scale_fill_gradient}}}
}
\description{
Plot the autocovariance of the residuals over the functional domain (e.g. time) for 
a function-on-scalar model fitted with \code{\link[refund]{pffr}}.
}
