% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plots.R
\name{plot_residsVSxy}
\alias{plot_residsVSxy}
\title{Residual plot vs space for \code{\link[refund]{pffr}} models}
\usage{
plot_residsVSxy(model, dat_xy, breaks, labels_cut = NULL, base_size = 11,
  legend_rows = NULL, legend_cols = NULL, mark_location = NULL,
  mark_location_size = 5, ...)
}
\arguments{
\item{model}{Function-on-scalar model fitted with \code{\link[refund]{pffr}}}

\item{dat_xy}{Data set precalculated with \code{\link{prepareData_residsVSxy}},
containing information on the mean residuals per measurement location.}

\item{breaks}{breaks to be used for the colors of the residual categories.
E.g. \code{c(-10,0,10)} to show the residuals categorized in \code{c("(-10,0]","(0,10]")}}

\item{labels_cut}{Optional labels for the categories defined using \code{breaks}.}

\item{base_size}{see \code{\link[ggplot2]{theme_bw}}}

\item{legend_rows, legend_cols}{Optional parameters specifying the grid of
legend elements in \code{\link[ggplot2]{guide_legend}}}

\item{mark_location}{Optional row number \code{dat_xy} for which the location
should be marked by a special black dot in the plot}

\item{mark_location_size}{Size of the black dot for \code{mark_location}}

\item{...}{Further arguments passed to \code{\link[ggplot2]{theme}}}
}
\description{
Plot the residuals of a function-on-scalar model fitted with 
\code{\link[refund]{pffr}}, which includes a spatial component, against the latter.
Note: This function can only be used for situations, where measurements
took place at a set of fixed locations in 2D, which all have multiple measurements.
Situations where each observations has a unique metric value on the x and/or y 
axis are not supported. Neither are other spatial dimensions than 2D.
}
