% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plots.R
\name{plot_resVSfitted}
\alias{plot_resVSfitted}
\title{Residual plot vs fitted values for \code{\link[refund]{pffr}} models}
\usage{
plot_resVSfitted(model, type = "response", base_size = 11,
  xlab = "fitted values", ylab = NULL,
  main = "Residuals vs fitted values\\nHeatmap of binned points",
  legend_limits = NULL, legend_breaks = waiver(), ...)
}
\arguments{
\item{model}{Function-on-scalar model fitted with \code{\link[refund]{pffr}}}

\item{type}{See \code{\link[mgcv]{residuals.gam}}}

\item{base_size}{Size of plot elements. see \code{\link[ggplot2]{theme_bw}}.
Defaults to 11.}

\item{xlab, ylab, main}{Optional plot annotation}

\item{legend_limits, legend_breaks}{Optional arguments, passed as \code{limits}
and \code{breaks} to \code{\link[ggplot2]{scale_fill_gradientn}}.}

\item{...}{Further arguments passed to \code{\link[ggplot2]{theme}}}
}
\description{
Plot the residuals of a function-on-scalar model fitted with 
\code{\link[refund]{pffr}} against its fitted values.
}
