% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plots.R
\name{plot_resVSyindex}
\alias{plot_resVSyindex}
\title{Residual plot vs functional domain for \code{\link[refund]{pffr}} models}
\usage{
plot_resVSyindex(model, base_size = 11, xlab = "yindex",
  ylab = "residuals",
  main = "Residuals vs yindex\\nHeatmap of binned points",
  legend_limits = NULL, legend_breaks = waiver(), ...)
}
\arguments{
\item{model}{Function-on-scalar model fitted with \code{\link[refund]{pffr}}}

\item{base_size}{Size of plot elements. see \code{\link[ggplot2]{theme_bw}}.
Defaults to 11.}

\item{xlab}{Optional plot annotation}

\item{ylab}{Optional plot annotation}

\item{main}{Optional plot annotation}

\item{legend_limits}{Optional arguments, passed as \code{limits}
and \code{breaks} to \code{\link[ggplot2]{scale_fill_gradientn}}.}

\item{legend_breaks}{Optional arguments, passed as \code{limits}
and \code{breaks} to \code{\link[ggplot2]{scale_fill_gradientn}}.}

\item{...}{Further arguments passed to \code{\link[ggplot2]{theme}}}
}
\description{
Plot the residuals of a function-on-scalar model fitted with 
\code{\link[refund]{pffr}} against its functional domain (e.g. time).
}
