% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plots.R
\name{prepareData_residsVSxy}
\alias{prepareData_residsVSxy}
\title{Data preparation for \code{\link{plot_residsVSxy}}}
\usage{
prepareData_residsVSxy(model, data, yvar, xCoord_var, yCoord_var, ...)
}
\arguments{
\item{model}{Function-on-scalar model fitted with \code{\link[refund]{pffr}}}

\item{data}{Data for which the residuals of the model should be looked at}

\item{yvar}{name of the response variable in \code{model}}

\item{xCoord_var, yCoord_var}{name of the variables containing location information
for the x and y axis of the measurements}

\item{...}{Additional arguments passed to \code{\link[refund]{predict.pffr}}}
}
\description{
Prepare the data set containing the mean residuals per measurement location
for the following call of \code{\link{plot_residsVSxy}}.
Note: This function can only be used for situations, where measurements
took place at a set of fixed locations in 2D, which all have multiple measurements.
Situations where each observations has a unique metric value on the x and/or y 
axis are not supported. Neither are other spatial dimensions than 2D.
}
