% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_CIs}
\alias{sample_CIs}
\title{Sample object created with \code{\link[FoSIntro]{calc_bootstrapCIs}}}
\format{An object of class \code{list} of length 7.}
\usage{
sample_CIs
}
\description{
Sample object created with \code{\link[FoSIntro]{calc_bootstrapCIs}}
containing the information for pointwise confidence intervals based
on a nonparametric bootstrapped with 1000 bootsrap samples. The object
is used for the creation of the package vignette.
}
\keyword{datasets}
