% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Earthquake data set used in the Bauer (2018) paper}
\format{An object of class \code{data.frame} with 15129 rows and 11 columns.}
\usage{
sample_data
}
\description{
This data set is the basis for the running example in Bauer (2018) and
contains data regarding simulated earthquakes based on a real event that
took place in 1994 in Northridge (US).
}
\keyword{datasets}
