% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_CIs.R
\name{simulate_gamma}
\alias{simulate_gamma}
\title{Function to draw one parametric bootstrap sample for a Gamma model}
\usage{
simulate_gamma(model)
}
\arguments{
\item{model}{Function-on-scalar model fitted with \code{\link[refund]{pffr}}}
}
\description{
This function draws one parametric bootstrap sample from a \code{\link[refund]{pffr}}
model using a Gamma response distribution.
}
