% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_CIs.R
\name{simulate_pffr}
\alias{simulate_pffr}
\title{Simulate new observations from a \code{pffr} model}
\usage{
simulate_pffr(model, sim_fun = "simulate", y_minValue = NULL)
}
\arguments{
\item{model}{Function-on-scalar model fitted with \code{\link[refund]{pffr}}}

\item{sim_fun}{Function that should be used to draw the parametric bootstrap
sample based on \code{model}. Default value is \code{'simulate'} which uses
the \code{\link[stats]{simulate}} function. For distribution families not
supported in \code{\link[stats]{simulate}} one can either use \code{sim_fun = 'gamma'}
to perform bootstrapping for Gamma models or one can pass the name of its own function
to perform the bootstrap. The function should thereby take a \code{pffr} model
as single argument and return a bootstrapped matrix of y values with one row
per observation and one column per functional index (e.g. time point), equivalent to
\code{\link[FoSIntro]{simulate_gamma}}.}

\item{y_minValue}{Optional argument. If specified, possible bootstrapped response
values smaller than \code{y_minValue} are set to \code{y_minValue}. This can e.g.
be used for Gamma models to prevent estimation problems of models based on the
bootstrapped data, caused by response values being too close to zero.}
}
\value{
Matrix of reponse values with one row per observation and one column
per functional index (e.g. time point).
}
\description{
This function can be used to draw a parametric bootstrap sample from a
function-on-scalar model fitted with \code{\link[refund]{pffr}}. Per default,
new observations are sampled using \code{\link[stats]{simulate}}. If the
distribution family used is not supported by \code{simulate}, one can also
define and pass an own function as \code{sim_fun}. One example for this is
\code{\link[FoSIntro]{simulate_gamma}}, which supports simulation from a
Gamma distribution and can be called using \code{sim_fun = "gamma"}.
}
