% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_CIs.R
\name{calc_singleBootstrapCI}
\alias{calc_singleBootstrapCI}
\title{Calculate a pointwise confidence interval band for a single smooth term in a \code{\link[refund]{pffr}} model}
\usage{
calc_singleBootstrapCI(bs_results, select, alpha = 0.05)
}
\arguments{
\item{bs_results}{Output of function \code{\link{bootstrap_pffr}}}

\item{select}{Index of smooth term to be looked at}

\item{alpha}{\code{(1-alpha)} CIs are calculated. The default 0.05 leads to 95\% CIs}
}
\description{
Function to calculate a pointwise confidence interval (CI) band for a single
smooth term, based on the output of \code{\link{bootstrap_pffr}}.
NOTE: Only 1D and 2D smooths are currently handled correctly!
}
