% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plots.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Plot model predictions for \code{\link[refund]{pffr}} models}
\usage{
plot_predictions(model, newdata, ci_type = "none", labels = NULL,
  legend_title = "group", col_palette = "Set1", col_vector = NULL,
  rev_cols = FALSE, base_size = 11, legend_symbol_size = 2,
  type = "response", log10 = FALSE, plot.margin = unit(rep(5.5, 4),
  "points"), hline, ybreaks = waiver(), ylabels = waiver(),
  xlab = "yindex", ylab = "prediction", main, ylim = NULL, lwd = 2)
}
\arguments{
\item{model}{Function-on-scalar model fitted with \code{\link[refund]{pffr}}}

\item{newdata}{See \code{\link[refund]{predict.pffr}}}

\item{ci_type}{One of \code{"none"} (default; no confidence bands), \code{"ci"}
(pointwise confidence intervals for the predicted mean) or \code{"pi"} (pointwise
prediction intervals), specifying which type of confidence bands should be plotted}

\item{labels}{Labels vector with length equal to \code{nrow(newdata)}.
Used for the legend.}

\item{legend_title}{Title of legend}

\item{col_palette}{Color palette to use in plot, see
\code{\link[ggplot2]{scale_color_brewer}}. Only used if \code{nrow(newdata)>1}
and if \code{is.null(col_vector)}.}

\item{col_vector}{Vector of colors to use}

\item{rev_cols}{If TRUE, the color palette is turned around.}

\item{base_size}{Base size of plot elements, see \code{\link[ggplot2]{theme_bw}}.
Defaults to 11.}

\item{legend_symbol_size}{Size of symbols in legend. Default to 2}

\item{type}{\code{type} argument for \code{\link[refund]{predict.pffr}}}

\item{log10}{If TRUE the predictions are shown on a log10 scale}

\item{plot.margin}{Margin of the ggplot}

\item{hline}{Vector of \code{\link[ggplot2]{geom_hline}}}

\item{ybreaks}{Numeric vector of breaks for the y axis}

\item{ylabels}{Character vector of labels for the y axis. Only used if
\code{ybreaks} is specified}

\item{xlab, ylab, main, ylim, lwd}{Further arguments for the plot function}
}
\description{
Plots the model predictions for a function-on-scalar model fitted with
\code{\link[refund]{pffr}}.
}
