% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plots.R
\name{plot_predVSobs}
\alias{plot_predVSobs}
\title{Plot predictions vs real observations for \code{\link[refund]{pffr}} models}
\usage{
plot_predVSobs(model, data, yvar, yvar_label, type = "response",
  log10 = FALSE, pi_type = "none", pi_alpha = 0.05, base_size = 11,
  xlab = "yindex", ylab, main = "Prediction vs observation", ybreaks,
  ylabels, ylim = NULL, lwd = 1, legend_symbol_size = 2, ...)
}
\arguments{
\item{model}{Function-on-scalar model fitted with \code{\link[refund]{pffr}}}

\item{data}{Dataset with one row for which the predictions are to be plotted}

\item{yvar}{name of the response variable in \code{model}}

\item{yvar_label}{Optional label of the response variable used for plot annotation}

\item{type}{\code{"response"} (default) or \code{"link"}.
See \code{\link[refund]{predict.pffr}}.}

\item{log10}{If TRUE the predictions are shown on a log10 scale}

\item{pi_type}{One of \code{"none"} (default; no uncertainty bands), \code{"pi_norm"}
or \code{"pi_gamma"} (pointwise prediction intervals), specifying which type of uncertainty bands should be plotted.
Prediction intervals are either calculated summing the predicted mean variance and
the residual variance (\code{"pi_norm"}) or using 'posterior simulation' of the parameter vector
(\code{"pi_gamma"}) as outlined here: https://stat.ethz.ch/pipermail/r-help/2011-April/275632.html.
Note that the \code{"pi_norm"} PIs are not generally applicable for generalized models.
The posterior simulation approach is more generally applicable, but is currently only implemented
for Gamma regression.}

\item{pi_alpha}{If \code{pi_type != "none"}, \code{1-pi_alpha}\% prediction
intervals are plotted. Defaults to 0.05.}

\item{base_size}{Base size of plot elements, see \code{\link[ggplot2]{theme_bw}}.
Defaults to 11.}

\item{xlab, ylab, main}{Optional arguments for plot annotation}

\item{ybreaks, ylabels}{Optional specification of the tick position and the
corresponding labels (character vector) for the y axis}

\item{ylim}{Optional y axis limits of plot}

\item{lwd}{Optional line width. Defaults to 2}

\item{legend_symbol_size}{Optional size of symbols in legend.}

\item{seed}{Additional arguments passed to \code{\link{pi_gamma}}}
}
\description{
Plots the predictions o a function-on-scalar model fitted with
\code{\link[refund]{pffr}} agains the real observations.
}
